Improve the way dependencies are declared
@jreese had some suggestions in my previous pull request after it got
merged, so this addresses some of those suggestions:

- Uses the PEP 508 `python_version` environment marker instead of
  conditional logic inside `setup.py`. I've updated `requirements.txt`
  to use this too.
- Split dev requirements into a separate `requirements-dev.txt`, and
  updated the README to include instructions for it.

This PR does not use pyup, because it looks like it's free for
non-commercial use only (I don't know that that means in this context),
and because this project isn't public yet.

It also does not include a makefile yet, because Jennifer and I need to
talk through where we'd stick the virtualenv in that case.

I tested these changes on 3.6 and 3.7.