Fix whitespace handling around strings
Strings are safe to use with a word operator in many cases, so this
fixes it to support that.

This was causing some parser errors in
`commerce/shopping_recommendations_service_shots.py` due to some whacky
f-strings. We weren't encountering those errors before, because of other
f-string parsing issues that we were hitting first.