Loosen types on deep_equals.
As written, this makes it a type-checker error if you pass in a value of
other that doesn't match the type of the node you're checking against.
If you want to ask a generic question similar to 'Does this equal
Node("True")', you can't use this function. It calls into a helper that
allows for (and checks as a base case) different types to be passed in,
so lets loosen types here to make this useful.