Change FormattedString's end field to a Literal
I noticed this while working on the documentation. I think this predated
pyre's Literal support.

I only did this for `end` and not `start`, because there's a large
combinatorial set of valid `start` values. If we wanted to better type
`start`, we should probably first separate the prefix from the opening
quotes.