Clarify LibCST license.
Make explicit in the LICENSE and README files the exact files that are not licensed under MIT since being forked. These files already have license information in their headers, but it is always good to be explicit at the top level as well. Also include a copy of the licenses in the LICENSE file to conform with requirements.