Count the information we add in ApplyTypeAnnotationsVisitor
Refactor ApplyTypeAnnotationsVisitor so that all annotation
information is added via a smart constructor method starting
with `_apply_annotation_to`. This makes it much easier to
skim the code and understand where annotations are actually
added with a simple forward search.

Then, add an AnnotationCounts dataclass and count up all the
annotations we add inside the transform. This should be helpful
for a few reasons:
- First, it just makes counting the annotations easier. Prior
  to this change, we would have to run some separate command
  to count annotations before and after a codemod, which is
  not as convenient as doing it directly, and would also fail
  to account for cases where we changed an annotation.
- Second, I want to be able to avoid altering the import
  statements in cases where we never actually made any changes.
  Having annotation counts will help us do this - we can just
  return the original tree (without import changes) in that
  situtation.

```
> python -m unittest libcst.codemod.visitors.tests.test_apply_type_annotations.TestApplyAnnotationsVisitor
................................................
----------------------------------------------------------------------
Ran 48 tests in 1.773s

OK
```

(