Add initial codegen and inflate support
The current codegen doesn't have logic for when to automatically
add parentheses yet.

It's possible that my inflate doesn't quite work right - I noticed
that the `ClassDef` inflate transfers some whitespace from parenthesis
tokens to the parent. But I don't think that's necessary here, I did
some manual testing in ipython and it seems to me that the whitespace
before `(` and after `)` always get assigned to the `with` and `:`
tokens, as I would have hoped, which means inflate doesn't need to care.