Clean up abstractions around annotations.
There were three different layers of target things, which
meant using target / targets as a naming convention was not
working well.

Now, I say:
- a target is one or more `target` values
- a target is a single lvalue, and maps to a `bindings` object
  of one or more bindings
- a binding is a specific, single value that can be annotated

Similarly, there are two layers in annotation handling:
- a type comment corresponds to an annotations, one or more types
- an annotations atom is an annotation

When zipping, we zip a single bindings against a single annotations.

When handling multi-assignment Assigns, we have to do this multiple
times and concatenate the results.