Attach a supplied Path when parsing a module
This adds an optional `path` parameter to `parse_module`, and then
attaches that path object to the final `Module` object after parsing.

This allows frameworks like µsort or Fixit to supply the path being
parsed, so that the path can be later accessed as a property of the
module, without needing to track or pass that through a stack via
alternative means. Visitors can then fetch this path information
from the root of the tree when needed, rather than needing separate
metadata providers or state information.